PerfProfiler: Fix signedness mismatch on comparison
Detected by GCC9.

+++ src/plugins/perfprofiler/perfprofilertracemanager.cpp
@@ -399,13 +399,13 @@ int PerfProfilerEventTypeStorage::append(Timeline::TraceEventType &&type)
     if (perfType.isLocation()) {
         const size_t index = m_locations.size();
         m_locations.push_back(perfType);
-        QTC_ASSERT(index <= std::numeric_limits<int>::max(),
+        QTC_ASSERT(index <= size_t(std::numeric_limits<int>::max()),
                    return std::numeric_limits<int>::max());
         return static_cast<int>(index);
     } else if (perfType.isAttribute()) {
         const size_t index = m_attributes.size();
         m_attributes.push_back(perfType);
-        QTC_ASSERT(index <= std::numeric_limits<int>::max(),
+        QTC_ASSERT(index <= size_t(std::numeric_limits<int>::max()),
                    return -std::numeric_limits<int>::max());
         return -static_cast<int>(index);
     }
@@ -416,7 +416,8 @@ int PerfProfilerEventTypeStorage::append(Timeline::TraceEventType &&type)
 int PerfProfilerEventTypeStorage::size() const
 {
     const size_t result = m_attributes.size() + m_locations.size();
-    QTC_ASSERT(result <= std::numeric_limits<int>::max(), return std::numeric_limits<int>::max());
+    QTC_ASSERT(result <= size_t(std::numeric_limits<int>::max()),
+               return std::numeric_limits<int>::max());
     return static_cast<int>(result);
 }
